refactoring packages and updating --all flag
